backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3        
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
deprecation==2.1.0       
dnspython==2.7.0
email-validator==2.3.0
fastapi==0.116.1
git-filter-repo==2.47.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
packaging==25.0
postgrest==1.1.1
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
realtime==2.7.0
requests==2.32.5
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
storage3==0.12.1
StrEnum==0.4.15
supabase==2.18.1
supabase_auth==2.12.3
supabase_functions==0.10.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
google-auth-oauthlib==1.2.1
google-api-python-client==2.154.0
google-auth==2.38.0


# YT-Agent-Umar Dependencies
openai==1.58.1
loguru==0.7.2
tenacity==8.2.3
redis==7.1.0
ffmpeg-python==0.2.0
